3 Examples
Descriptions are provided in the p5.js code of each example.
1. MarechalLabAssignmentSeven_2025_03_31_13_53_21.zip
Three stars rotate in the middle of a canvas. Their color is random every time the code is run. This color randomization is added as an aesthetic change to give a feel of personalization. This is inspired partially by Flappy Bird, a retro-like game, which can dish out randomly colored backgrounds and bird colors. FUN FACT: The odds of the same exact combination of 3 star colors appears barring position is 2^72. If position is considered, the number shoots up drastically to 2^216. For reference, the chances of winning the lottery are about 1 in 300 million, which is close to odds of 2^20, nowhere near as unlikely as the other two aforementioned scenarios. 
2. MarechalLabAssignmentSix_2025_03_17_11_23_50.zip
Six ghosts and four spiders move around a dark background. Spiders move at a speed based on their size. Wisps move erratically. By clicking, ghosts will follow the mouse. When ghosts move, they spread ectoplasm that make the background brighter. When spiders move, their steps make the background darker. Ghosts phase through the spiders since the spiders are living and the ghosts are dead.
3. MarechalLabAssignmentFive_2025_03_11_04_07_17.zip
Contained in this example is the code for a custom-made theme, which is used to resemble a TV screen. Six rectangles make up the main screen, and below it are 3 buttons decorated with care. The center button is a basic white and navy structure. When the mouse is in its bounds and the SPACEBAR is pressed, the big rectangle behind the others turns into a deep blue. To the right of that button is another button with a color theme surrounding smaragdine, a fancy shade of green. When the mouse is in the bounds of the button and then is pressed and released, the opacity of the hot purple and pink rectangles is inverted. To the right of that button is a button with a color theme surrounding amaranth, a fancy shade of red. When the mouse is clicked in the bounds of this button, the opacity of the red, dark orange, and orange rectangles is inverted. The stroke of the large backdrop rectangle gives the illusion of color change, when in actuality the coming of opacity in rectangles makes their color the same as the background color.